Rewrite API to Swagger 2.0
The sequence of reasons for this is the following. Swagger codegen 3.x
dropped support for several languages, most notably all JavaScript
clients. Swagger codegen 2.x, however, doesn't support Open API 3.0
format. So, in order to generate a JavaScript client, we have to use the
older codegen, which implies using the older format.